+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - `isLazyPromise(x)` becomes `x instanceof LazyPromise`.
 
+- Generator syntax.
+
 ## [@lazy-promise/core@0.0.16](https://github.com/lazy-promise/lazy-promise/tree/%40lazy-promise/core%400.0.16)
 
 - Make `log` play well with string substitutions.